refactor: move cli functions to api
