+++ junifer/api/__init__.py
@@ -5,6 +5,7 @@
 # License: AGPL
 
 from . import decorators
+from .functions import collect, list_elements, reset, run, queue
 
 
-__all__ = ["decorators"]
+__all__ = ["decorators", "collect", "queue", "run", "reset", "list_elements"]
