+++ junifer/api/tests/test_functions.py
@@ -1,4 +1,4 @@
-"""Provide tests for cli functions."""
+"""Provide tests for API functions."""
 
 # Authors: Federico Raimondo <f.raimondo@fz-juelich.de>
 #          Leonard Sasse <l.sasse@fz-juelich.de>
@@ -13,7 +13,7 @@ import pytest
 from ruamel.yaml import YAML
 
 import junifer.testing.registry  # noqa: F401
-from junifer.cli.functions import collect, list_elements, queue, reset, run
+from junifer.api import collect, list_elements, queue, reset, run
 from junifer.datagrabber.base import BaseDataGrabber
 from junifer.pipeline.registry import build
