+++ junifer/cli/__init__.py
@@ -4,7 +4,5 @@
 # License: AGPL
 
 from .cli import cli
-from .functions import collect, queue, run
 
-
-__all__ = ["cli", "collect", "queue", "run", "reset", "list_elements"]
+__all__ = ["cli"]
